requirements.txt
django
djangorestframework
pygments

sqlalchemy
pony
# pip install -r requirements.txt